Refactored code for better build
1. created setup.py
2. fix root drectory to dlio_benchmark
3. renamed dlio_benchmark.py to main.py
4. renamed dlio_postprocessor.py to postprocessor.py
5. fixed documentation to use dlio_benchmark and dlio_postprocessor entry points.
