+++ dlio_benchmark/utils/utility.py
@@ -26,12 +26,15 @@ from typing import Dict
 
 import numpy as np
 import inspect
-
+import psutil
+import socket
 # UTC timestamp format with microsecond precision
 from dlio_benchmark.common.enumerations import LoggerType
 
 LOG_TS_FORMAT = "%Y-%m-%dT%H:%M:%S.%f"
 from mpi4py import MPI
+
+p = psutil.Process()
 
 def add_padding(n, num_digits=None):
     str_out = str(n)
@@ -137,6 +140,8 @@ def create_dur_event(name, cat, ts, dur, args={}):
         tid = threading.get_ident()
     else:
         tid = 0
+    args["hostname"] = socket.gethostname()
+    args["cpu_affinity"] = p.cpu_affinity()
     d = {
         "name": name,
         "cat": cat,
