+++ docs/source/install.rst
@@ -6,9 +6,8 @@ DLIO itself should run directly after installing dependence python packages spec
 
     git clone https://github.com/argonne-lcf/dlio_benchmark
     cd dlio_benchmark/
-    pip install -r requirements.txt 
-    export PYTHONPATH=$PWD/:$PYTHONPATH
-    python ./dlio_benchmark/benchmark.py
+    pip install .
+    dlio_benchmark
     
 One can build docker image run DLIO inside a docker container.  
 
@@ -17,18 +16,18 @@ One can build docker image run DLIO inside a docker container.
     git clone https://github.com/argonne-lcf/dlio_benchmark
     cd dlio_benchmark/
     docker build -t dlio .
-    docker run -t dlio python ./dlio_benchmark/benchmark.py
+    docker run -t dlio dlio_benchmark
 
 A prebuilt docker image is available in docker hub 
 
 .. code-block:: bash 
 
     docker pull docker.io/zhenghh04/dlio:latest
-    docker run -t docker.io/zhenghh04/dlio:latest python ./dlio_benchmark/benchmark.py
+    docker run -t docker.io/zhenghh04/dlio:latest dlio_benchmark
 
 To run interactively in the docker container. 
 
 .. code-block:: bash
 
     docker run -t docker.io/zhenghh04/dlio:latest bash
-    root@30358dd47935:/workspace/dlio# python ./dlio_benchmark/benchmark.py
+    root@30358dd47935:/workspace/dlio# dlio_benchmark
