+++ tests/test_data/.hydra/hydra.yaml
@@ -16,11 +16,11 @@ hydra:
     footer: "Please submit questions/bugs to \n  https://github.com/argonne-lcf/dlio_benchmark/issues\n\
       \n          Copyright (c) 2021 UChicago Argonne, LLC"
     template: "\n${hydra.help.header}\n\nDLIO - an IO benchmark for deep learning\
-      \ applications. \n\nRunning the benchmark: python dlio_benchmark/benchmark.py workload=unet3d\n\
+      \ applications. \n\nRunning the benchmark: python dlio_benchmark/main.py workload=unet3d\n\
       \nOne can select the workload configuration using \"workload={WORKLOAD}\". \n\
       The corresponding YAML file is ./configs/workload/{WORKLOAD}.yaml folder. \n\
       Available choise for $APP_CONFIG_GROUPS\nOne can override everything in the\
-      \ command line, for example:\npython dlio_benchmark/benchmark.py workload.framework=tensorflow\n\
+      \ command line, for example:\npython dlio_benchmark/main.py workload.framework=tensorflow\n\
       \nOne can also create a custom YAML file for a specific workload. \nAn example\
       \ of a YAML file is as follows. \n\n-------\n$CONFIG\n-------\nA complete list\
       \ of config options in the YAML file can be found: \nhttps://argonne-lcf.github.io/dlio_benchmark/config.html\n\
